+++ release-notes.md
@@ -1,6 +1,12 @@
 ← [README](README.md)
 
 # Release notes
+## 2.1.0
+Released 29 June 2024.
+
+* The `I18n.GetByKey` method and `I18n.Keys` constants are now available by default.
+* Fixed error in rare cases when an `i18n/default.json` file is invalid.
+
 ## 2.0.1
 Released 08 May 2022.